fix(dsa): use attn_kv_dtype and skip invalid BF16 slots
- Builders take attn KV dtype from kv_cache_spec.dtype instead of
  global cache_config.cache_dtype (indexer KV stays independent).
- Rename builder/impl field to attn_kv_dtype and clarify indexer
  vs attention dtype docs.
- A5 BF16 format maps invalid slots to [-1,-1]; scatter skips them
  instead of clamping -1 to 0.

+++ tests/ut/attention/test_dsa_v1.py
@@ -59,7 +59,7 @@ def _make_builder(compressor_ratio: int = 4) -> AscendDSAMetadataBuilder:
         speculative_config=None,
         parallel_config=SimpleNamespace(tensor_parallel_size=2),
     )
-    kv_cache_spec = SimpleNamespace(compress_ratio=compressor_ratio, block_size=128)
+    kv_cache_spec = SimpleNamespace(compress_ratio=compressor_ratio, block_size=128, dtype=torch.bfloat16)
     with (
         patch(
             "vllm_ascend.attention.dsa_v1.kv_cache_dtype_str_to_dtype",
